# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/utilities": {
      "get": {
        "operationId": "UtilitiesController_list[0]_v1",
        "parameters": [
          {
            "name": "bu_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "utilities"
        ]
      },
      "post": {
        "operationId": "UtilitiesController_create[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUtilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "utilities"
        ]
      }
    },
    "/utilities": {
      "get": {
        "operationId": "UtilitiesController_list[1]",
        "parameters": [
          {
            "name": "bu_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "utilities"
        ]
      },
      "post": {
        "operationId": "UtilitiesController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUtilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "utilities"
        ]
      }
    },
    "/v1/utilities/{utilityId}": {
      "patch": {
        "operationId": "UtilitiesController_update[0]_v1",
        "parameters": [
          {
            "name": "utilityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUtilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "utilities"
        ]
      }
    },
    "/utilities/{utilityId}": {
      "patch": {
        "operationId": "UtilitiesController_update[1]",
        "parameters": [
          {
            "name": "utilityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUtilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "utilities"
        ]
      }
    },
...
    "/v1/users/sync": {
      "post": {
        "operationId": "UsersController_sync[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or update a user profile from Identity Federation (never credentials)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/sync": {
      "post": {
        "operationId": "UsersController_sync[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or update a user profile from Identity Federation (never credentials)",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/{itsId}": {
      "get": {
        "operationId": "UsersController_get[0]_v1",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "User profile with role assignments (filtered to your workspace)",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update[0]_v1",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/{itsId}": {
      "get": {
        "operationId": "UsersController_get[1]",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "User profile with role assignments (filtered to your workspace)",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update[1]",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/{itsId}/roles": {
      "get": {
        "operationId": "UsersController_roles[0]_v1",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/{itsId}/roles": {
      "get": {
        "operationId": "UsersController_roles[1]",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users": {
      "post": {
        "operationId": "UsersController_create[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/v1/user-roles": {
      "post": {
        "operationId": "UserRolesController_assign[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a role to a user in a scope (CORE | BUSINESS_UNIT bu_id | UTILITY utility_id)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserRolesController_revoke[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a role assignment",
        "tags": [
          "users"
        ]
      }
    },
    "/user-roles": {
      "post": {
        "operationId": "UserRolesController_assign[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a role to a user in a scope (CORE | BUSINESS_UNIT bu_id | UTILITY utility_id)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserRolesController_revoke[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a role assignment",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Miqaat Core Identity Authorization Service",
    "description": "Dynamic Business Unit / Utility / Application / Environment / Module / Role / Permission model, client registry and server-side authorization checks. No API keys: every call carries an RS256 bearer JWT verified via JWKS - service tokens (typ client-authentication+jwt) from registered service principals, or administrator access tokens (typ at+jwt) issued by Identity Federation. Never stores credentials.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256, verified via JWKS)",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUtilityDto": {
        "type": "object",
        "properties": {
          "bu_id": {
            "type": "string",
            "description": "Parent business unit"
          },
          "name": {
            "type": "string",
            "example": "Helpdesk"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ],
            "default": "active"
          }
        },
        "required": [
          "bu_id",
          "name"
        ]
      },
      "UpdateUtilityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
      "SyncUserDto": {
        "type": "object",
        "properties": {
          "its_id": {
            "type": "string",
            "example": "30337752"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ]
          }
        },
        "required": [
          "its_id"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "its_id": {
            "type": "string",
            "example": "31189012",
            "description": "ITS ID is the user ID"
          },
          "name": {
            "type": "string",
            "example": "Burhan"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ],
            "default": "active"
          }
        },
        "required": [
          "its_id",
          "name"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ]
          }
        }
      },
      "UserRoleDto": {
        "type": "object",
        "properties": {
          "its_id": {
            "type": "string",
            "example": "31267890"
          },
          "role_id": {
            "type": "string"
          },
          "scope_type": {
            "type": "string",
            "enum": [
              "CORE",
              "BUSINESS_UNIT",
              "UTILITY"
            ],
            "example": "UTILITY"
          },
          "scope_id": {
            "type": "object",
            "description": "NULL for CORE, else bu_id or utility_id",
            "nullable": true
          }
        },
        "required": [
          "its_id",
          "role_id",
          "scope_type"
        ]
      },
...
    }
  }
}